Adds backward compatibility hack to fix loading old TPState objects.
TPState objects that were serialized to disk prior to fixing the
basis serialization of TPState objects (in commit
aa8a29049a9c3ddafe598d903602ab5afdb7aad4) were unable to be
loaded because they were at least sometimes saved with an empty basis.
This issue is fixed by this commit by adding a HACK in tpstate.py that
checks for an empty basis and converts this to None.  This isn't a
complete patch, as the basis should really be something non-None, but
it's the best we can do given there's no information stored about the
basis.