Fix forced germ list issue with deduping
Patches a problem that arises sometimes when using a user-specified list of germs together with deduping. The code expects that the user specified list will be included in the list of available candidates, which after deduping isn't always the case. This can also fail if the settings used to construct the candidate germ set are incompatible with the properties of the forced germ set (either because they are too short or maybe only a random subset of a particular length were initially included etc.) This patches that be adding a check for whether all of the forced germs are in the candidate set and if not appends the requisite ones.