Fixes a bug in MongoDB serialization.
Fixes the implementation of the 'circuit-str-json' auxtype, which is
used to serialize fiducial pair dictionaries (in experiment designs).
The metadir implementation is fine, but when I implemented the MongoDB
version I didn't realize exactly what was being serialized (I thought
it was just a list of circuits that should be stored as a list of
JSON circuit string, but really it can be anything JSONable that
may contain Circuits and doesn't contain any string, and when
serialized the Circuit objects are converted to strings).  This
commit fixes this oversight so that now the 'circuit-str-json' auxtype
works properly with MongoDB serialization.

This commit also adds a backwards compatibility HACK for loading
in old objects that were written before the MongoDB serialization
code existed and thus don't have {... _dbcoordinates: 'none'} in their
auxtypes member.  The hack adds a _dbcoordinates: 'none' entry to
the auxtypes dictionaries of objects that don't have them.