Adds a safeguard for saving results objects given an already_written_data_id.
Adds a warning and recovery attempt for the case when a results object
is being saved, is given the ID of a data object that is supposedly
already saved, but the data object doesn't contain any ._dbcoordinates
(i.e. this attribute is None).  Previously the code would just error,
and now it attempts to save the data object.  This really shouldn't
be needed (hence the warning), but it seems like good insurance.