Fixes bug in MongoDB serialization of 'json' auxiliary type.
Removes a spurious pickling of 'json' data before saving it to
the database, mostly likely from a copy/paste error of the
'pickle' auxtype.  Before this fix, 'json' auxtype data would not
be saved correctly, and would be unloadable (eek!).  The prominent
example of a 'json'-auxtype member is the 'idealout_lists' attribute
of an RB experiment design.