Updates MongoDB serialization so ProtocolData children are generated before writing.
A ProtocolData object generates sub-data objects lazily, and when writing to
disk only saves generated sub-data objects.  This presents an issue for MongoDB
documents, however, which default to writing with "overwrite_existing=False".
If a data object is saved without it's sub-data and later saved with some/all
of them generated the user must awkwardly set "overwrite_existing=True" even
though the data object hasn't really changed.  As a (at least temporary) solution
to this, this 1-line-change commit causes all the sub-data objects to be generated
and then saved when a ProtocolData object is written to a MongoDB.  We could think
about more clever ways to do this if write performance becomes an issue, but
it's not clear at this point whether that will be an issue.