Adds a HACK to wildcard1d processing so it works for more model types.
Causes the wildcard1d post-processing, which requires a gauge-optimized
model, to use the 'stdgaugeopt' model within an Estimate if one is available,
and to only perform a default gauge-optimization when one isn't (see comment
in code for more details).  This avoids, in cases where there is a stdgaugeopt
model, failures within wildcard1d computation when the model cannot be gauge-optimized
using default gaugeopt_to_target arguments (e.g. in the case of a reduced model
that needs to be converted to TP to gauge-optimize).  In the future, something
could be plumbed to convey the name of a gauge optimized model to the wildcard1d
routines; this is a quick fix that should work most of the time.