Add support for initializing greedy search from scratch
This commit adds support for starting the greedy search from scratch as is required when the 'force' argument is set to None (as opposed to 'singletons' or a user-specified list. In this case we now iterate through the candidate germ list and select the best germ according the the composite score.