Minor logging fix
Previously the first outer iteration of the greedy search would mistakenly report that the initial germ set amplified only a single parameter (which is incorrect in general). This is because the number of amplified parameters is only available following the initial outer iteration. We could add in a calculation to get the correct number for the initial germ set at the start of the search, but this is computationally expensive, so I went with the cheap route and added a flag so that we simply don't report the number of amplified parameters until we've done at least one outer-iteration of the greedy search.