Add initial germ set AC test
Adds a test (along with a flag for disabling this for expediency) that checks the initially constructed germ set, either from the argument passed in for 'force' or as greedily selected when force is None, for amplificational completeness using the test_germs_list_completeness function. If it is already AC then we short-circuit the search and simply return the initial list and print logging information to that effect.