Add additional unit tests
Adds additional unit test covering the case where force=None and the initial germ set is selected greedily. Also adds a test covering the case where a germ in a user specified list of forced germs is outside the candidate set, either because it is outside of the specs (i.e. longer, not sampled randomly etc.) or because it may have been originally but dropped during the de-duping.