Revert back debug changes and manually build cython
This removes the debugging workflow that was added and reverts back the changes to main.yml that had temporarily skipped the full testing gamut on the main workflow. This adds a line to manually build the cython extensions. Not sure why this isn't being done properly by the pip install call, but regardless this seems to get the tests working again and modules importing correctly.