Patch wildcard gauge optimization bugs
This patch addresses a bug in the 1D wildcard models that resulted from a mismatch in the gauge optimization settings used in different places in the code. We now have some basic plumbing for the GSTGaugeOptSuite that pipes the different gauge optimizations through to the badfit estimate code and makes sure that we are consistent. Also makes the wildcard models switchable on the gauge optimization in the report generation for cases where we have more than one gauge optimization used.