Unit test warning clean up in algorithms
Adds ignore for common but non-problem warning in pytest conf
Trying to get ahead of numpy.matrix deprecation with todense -> toarray.
Note that full removal of this dependency relies on scipy.sparse updates,
which started in scipy 1.8 but are not mainstream yet.
Fixed a few other minor warnings in algorithms tests