Rework extra test workflow
Reworks the way the pytests were being called to revert back to using automatic test discovery coupled with the exclusion of particular directories. Also remove the matrixing of the individual test packages. This should save on overhead for build time and environment set up and tear down on the runners.