Add notebook regression tests
This commit adds in notebook regression tests into the test extras workflow (for time considerations so we aren't always running these, since I anticipate these will take a while to run). The regression testing is managed by the nbval extension for pytest running in 'lax' mode. This means the extension is only checking that the cells run to completion and is not checking for the correctness of the output (this is something we could do, but would require a bunch of legwork on our end to prepare for initially). A few of the notebooks include examples that either take a long time to run or aren't suited to running on github actions. This includes things like running 2-qubit GST, calculating 2-qubit hessians, and api calls to the IBMQ experience (I'm not sure if this last thing would work on github actions, but it seemed like a PITA). For these nbval allows for either including a comment or a cell tag to indicate cells it should skip. I opted for cell tags since they live in the metadata and won't clutter up the notebooks or add confusion to users. Unrelatedly I fixed a build environment issue for macos (I had forgotten to swap out the build script).