Refactor tests in test_packages
Pytest's test discovery looks like it is more strict about the naming convention used for test files than nose was. We (either intentionally or incidentally) happened to have the correct convention consistently used in the main unit testing module, which is why we didn't notice this earlier. The names of the files have been updated in accordance with the pytest guidelines. Also fixes a typo in the call to the notebook regression tests.