Initial changes for additional support of CircuitListsDesign reporting
A number of the reporting functions for generating figures for reports presently have hardcoded assumptions that the experiment has a PlaquetteGridCircuitStructure associated with it. More general CircuitListsDesign based experiments don't have this sort of structure associated with them. Certain plotting functionality clearly relies on the existence of a plaquette structure, but others like per-sequence model violation histograms and related figures don't. This is an initial round of patches that aim to add more robust support for producing reports for experiments based on CircuitListsDesigns. This adds support for the per-sequence model violation histogram and scatter plots.