Better ModelTest support for CircuitListsDesigns and lack of gauge optimization
The reports for the results of a ModelTest didn't always produce things like the error generators for a model if there wasn't gauge optimization performed. This adds more robust reporting support for ModelTest results where we don't perform gauge optimization so that the error generators of the non-gauge-optimized model get reported in that case. This is done by adding a trivial gauge optimized model to the estimate that is simply a copy of the model being tested.