Fixes collected-parameter, parameter label, and parameter bounds serialization in models.
Models with collected parameters would not be serialized correctly because when
constructing the member objects (when de-serializing) we'd call set_gpindices with
parent=None.  This would rightly cause _clean_paramvec, when it was called later on,
to treat all the set gpindices as invalid and just allocate new indices.  This commit
adds a `parent_model` argument to ModelMember.from_memoized_dict, allowing for correct
setting of the parent model when members are intialized.

This commit also adds serialization of the parameter labels and parameter bounds
attributes held by a Model, which previously weren't serialized at all.  I don't
think anyone needs these right now, but it seems like a good thing to serialize them :)