Fixes a report generation bug dealing with wildcard budget objects.
Fixes the initialization of the switchboard's wildcard_budget_optional
attribute so that when supplied with a *dictionary* of WildcardBudget
objects wildcard_budget_optional gets set to the appropriate element of
the dictionary (the once corresponding to the current gauge-opt key) rather
than the entire dictionary.  Prior behavior would result in an error when
trying to construct a report summary-tab table.