Improves ProtocolResultsDir read-in from mongodb.
Previously, a ProtocolResultsDir would always read in *all* the results
in the database corresponding to its ProtocolData.  This can run into
trouble when we update results, e.g., by re-running an existing analysis
after tweaking or fixing it, since then there will be multiple results
in the db with the same key (the old results don't get deleted until
we explicitly delete them).  This commit updates the loading so that
only the results properly associated to the results directory (when
it was last saved) are loaded again, and adds a `read_all_results_for_data`
flag that can be set to regain essentially the old behavior (it's
actually a bit better than the old behavior since it will rename
duplicates so they're all loaded instead of just clobbering them
and loading the last one).