Adds listtools.lists_to_tuples and uses it when reading serialized parameter labels.
Parameter labels, particularly at the model level, tend to be tuples like
(gate_label, local_param_name) and when serialized and read back in these
result in unhashable lists which can be problematic.  The fact that Label
objects appear and really should be serialized as objects we'll leave for
another time - this commit just does the minimal work of converting lists
back into tuples upon reading the serialized parameter labels of a model.

This commit also adds a lighter version of this behavior for model members
themselves, which just converts top-level lists into tuples, in case a
model member object uses tuples in its local labels.