Fixes bug in Circuit.serialize() and adds expand_subcircuits argument.
Circuit.serialize was missing a repetition count when serializing a circuit,
which can result in the output being incorrect.  As a convenience, a new
`expand_subcircuits` argument is added that expands the subcircuits in a
Circuit object before serializing it, essentially serialzing the subcircuits
within the circuit too -- to this is kind of like a "recursive" option.