Adds a fix to MongoDB serialization.
Temporarily fixes a case where 'dbcoordinates' gets added to the
MongoDB documents associated with a NicelySerializable pygsti object.
The issue, as documented in a comment as well, is that to_nice_serialization()
can't tell whether it's being called to serialize to a MongoDB database or
to something else (like disk files), and it's supposed to add a 'dbcoordinates'
state attribute only in the latter case.  The fix uses `inspect` to check
whether it's being called from a mongodb function, but this is fragile and
should be replaced with something better (like a new argument?) in the future.