Updates backslash-containing strings in Circuit._write_q_circuit_tex
A number of string literals in Circuit._write_q_circuit_tex used unescaped
backslashes (!).  I think this would cause this method to fail to produce
the intended output (but it was probably never used), as Python would read
the backslash as part of an escape sequence.  In any case, this commit fixes
such string by either escaping the backslash or promoting the string to a
raw string.