IBM Q provider API updates/deprecation fixes
IBMQ has migrated to a new API for set up providers and using backends for running jobs on their hardware. The most pertinent change is in the objects supported by the backend.run() method. Previously this accepted Qobj objects which were containers for sets of circuits and meta information on the runtime parameters to use when running a job (i.e. number of shots etc.). Now this function takes as input a list of QuantumCircuit objects (or a single one) and the runtime parameters as kwargs. Should be compatible with both the old and new APIs while the old API until the old API is eventually deleted.