Add functionality for precomputing Jacobians
Adds functionality for precomputing the jacobians used in the per-germ global FPR scheme. I want to be able to use MPI for these while using native BLAS multithreading for other parts of the codebase, so adding this option for splitting this off looked to me to be the easiest way to able to have my cake and eat it too here.