Reimplement fisher information by L to address bug
This commit reimplements the calculation for the fisher information by L to address some buggy behavior observed in the plaquette based one wherein certain circuits that were indeed in the experiment design were nonetheless not appearing in any of the plaquettes or in additional_circuits. This itself is something we should track down, but for now this reimplements the calculation using circuit lists instead of the PlaquetteGridCircuitStructure approach. (it is incidentally much simpler too).