Patch MPI message size problem
The fisher information matrices for 3-qubits are often large enough to exceed the default ~2Gb message buffer limit for MPI. I could figure out a chunking scheme to work around this, but for now we actually only need the fisher information on the base rank, so we can skip this broadcast for the time being.

This commit also adds some more progress logging the FIM term cache construction.