Updates OpModel.copy() to preserve parameter bounds.
Adds logic to .copy() when a model contains both parameter bounds and
parameter labels to rebuild the parameter vector immediately (with copy())
and copy over the parameter bounds, thus preserving them.  I don't think
there's a guarantee that the newly created parameter vector in the copy
will be ordered the same as the original, so there's an assertion that
the parameter label ordering doesn't chance to avoid mis-assigning the
parameter bounds.