Patch a bug related to _remove_spam_vectors
The bug we were hitting in the interpolated ops notebook was related to a function used in the germ selection code that deletes spam vectors from a model. Marking the model for rebuild patches the issues. No idea why this broke now but not earlier (any why it seems inconsistent across python environments when we do).