Germ selection static gates
The greedy low-rank update implementation of germ selection ran into issues when encountering germs consisting of entirely parameter-less gates as this results in all zero jacobians (which the other implementations were able to handle gracefully but this implementation didn't). Of course, we don't really want to consider these types of germs in the first place as they are by definition insensitive to everything, so this adds logic to filter these out. Not perfect as this currently only works for explicit models (but then so does germ selection in general).