Front load COPA layout construction for memory checks
This shifts the construction of the COPA layouts for each iteration of a GST to before the actual optimization starts by using the ability of MDC stores to take a precomputed layout on construction. There are memory checks that happen at runtime when the layouts are constructed (which are a bit tricky to split off on their own since they depend heavily on the runtime environment) and this should make it easier to catch memory shortages before we get too deep into a run.