Add CPTP model support for choi eigenvalue function
It looks like the choi eigenvalue related functions assume that a gate has access to a dense operator interface, which isn't true in general for CPTP model which derive instead from LinearOperators. A simple patch is to simply call to_dense (which most linear operator derivatives do have) since we only need the transfer matrix in the end.