Add more general for seed model selection to StandardGST
Add more generalized support for setting the initial model used as the seed for each of the modes in StandardGST. Can now pass in some combination of Model  , GSTInitialModel or a dictionary of these to specify the desired behavior.

This also supersedes some advanced behavior that was accessibly by manually specifying a value for the starting_point attribute of StandardGST, so that old (probably unused) interface has been removed.