Add truncation tolerance for conversion to CPTP models
Add a new kwarg for convert for operation modelmembers that sets a tolerance for how much of a change (as measured by frobenius distance) we are willing to accept between an initial process matrix and it's CPTP truncated version before throwing an error.

There was already an assertion in place here, so the main change is adding a handle to what was previously a hard coded tolerance and then plumbing that through to the relevant functions.