Adds stopgap in report generation to ignore KeyError when creating Gram matrix plot.
Adds a try/except to gauge.py that catchs a KeyError when creating the
report GramMatrixPlot and just creates a BlankTable in its place.  This
error can occur when the experiment design doesn't have a complete set of
LGST circuits, which include the ones used for the Gram matrix plot.  In
the future, a more robust solution for ignoring/removing errors during
report generation should be employed.