Update requirements and add RTD config file
This commit updates the minimum version for sphinx used for building the documentation. It also adds a yaml file for the RTD configuration settings in accordance with new requirements from them. They have deprecated the ability to configure build settings through the online dashboard and will disable builds w/o this file in the fall.