Switch from cirq to cirq-core in setup.py
This commit switches the cirq dependency to cirq-core from cirq. Recently we've begun having pip installation issues due to unresolvable dependency conflicts. I traced this back to cirq, and while I haven't definitely proven this think that it is specifically a dependency issue with the cirq-rigetti submodule. When installing cirq it will by default include all of the optional submodules, most of which are specific hardware platform interface code which isn't needed for our purposes (which is just the ability to output cirq circuit objects). So we should be fine switching to cirq-core and avoiding installation of all of the extra submodules altogether. If a user wants that they can figure out the dependency issues for whichever hardware backend they want to connect to with cirq on their end.