Refactor do_idle_tomography
Refactor do_idle_tomography to remove the separate jacobian calculation option. Also swaps out instances of affine with active, where appropriate. Default behavior is to include all of the error generator terms in the Jacobian calculation and the options for not including some terms has been relegated to the advanced options dict.