Adds partial support for re-setting the state space of a modelmember.
Adds a setter method for the `ModelMember.state_space` property which calls
the new method `_update_submember_state_spaces`.  This new method's job is
to update as necessary the state space of any submembers and auxiliary data
(i.e. other than the _state_space attribute itself) needed to change the state
space in the object, such as re-creating representation objects.

This commit adds implementation for composed and embedded ops, and dense
ops.  Other op types may need specific implementation, but these are the
main two types needed for changing the state spaces of many-qubit models.
For example, it may be appropriate for dense ops to throw an error if the
state space changes in size since this change cannot readily be done.

Also, the _update_submember_state_spaces function should probably use
a memo, following the pattern of other functions that operate on webs
of modelmembers, so that modelmembers don't get asked to update their
state spaces many times and get stuck in an infinite update loop.

All that said, the current implementation allows changes of state
space in simple many-qubit models.