Alters behavior of EmbeddingOpFactory to pass sslbls through to embedded factory.
Previously, when an EmbeddingOpFactory was called on to create an operation for
a set of state space labels (sslbls) it would *not* pass these labels to the
contained (embedded) factory.  The thinking (if I remember correctly) was that this
information wouldn't be needed by the contained factory since the EmbeddingOpFactory
would handle all of the embedding.  While this latter statement is true the contained
factory *does* need the sslbls if its job is to create different operations on different
target qubits - even though the operation being constructed is a local (e.g. 2-qubit out
of the N total qubits) one.

Previously in the call to EmbeddedOpFactory.create_op `None` as passed down to the contained
factory's create_op call instead of the original `sslbls`.  This commit updates this so
that `sslbls` is passed through.

This means that in UnitaryOpFactory and in user-defined factories the `sslbls` parameter
of create_op will no longer be `None` all the time (specifically when such factories are
used with EmbeddingOpFactory objects) and so assertions that this be the case have been
removed from UnitaryOpFactory and the user-defined factory in the OperationFactories.ipynb
tutorial notebook.