Initial implementation of checkpointing for GST
Initial implementation of checkpointing for GST. This initially implementation targets the GateSetTomography protocol specifically and future updates will add support for additional protocols. To support the checkpointing functionality a new generator function version of the iterative GST algorithm has been implemented to enable lazy evaluation. This commit also introduces a new ProtcolCheckpoint parent class and the GateSetTomographyCheckpoint class. Each protocol requires a somewhat different set of information to enable it to checkpoint properly during runtime, so the current vision is that each protocol will have a specialized checkpoint object associated with it.