Initial implementation of StandardGST checkpointing
This commit adds an initial implementation of the StandardGSTCheckpointing object and it's use in the StandardGST protocol. Also includes a few bug fixes for the implementations of GateSetTomographyCheckpoint and ModelTestCheckpoint. The idea behind the implementation of the StandardGSTCheckpoint is to treat it as a container for some number of checkpoint objects for each of the sub-protocols run during StandardGST. To facilitate this I have also added a 'parent' attribute to the base ProtocolCheckpoint object to enable a link between child objects and their parents. This is primarily used at present to swap the child's write method with that of the parent's (which invokes the parent's serialization instead) when a parent checkpoint is present. Bugfixes were related to a mistake in using mutable objects as default kwarg values.