Deprecate use of comma-separated string for modes
The old behavior for the modes argument is kind of wonky and doesn't have any analogues in the code that I can recall. This switches the argument to take either an iterable or a single string (i.e. for specifying a single-mode), and adds a deprecation warning about the comma-separated version.