Add new RB Unit Tests
This commit adds new unit tests targeting the construction of BiRB experiment designs as well as running the protocol and some basic checks for correct output in the ideal case. This commit also plugs holes in the previously existing set of unit tests for the remaining RB protocols (CRB, DRB, MRB) and adds tests for running those protocols (previously only the experiment design creation was being tested for these protocols). Finally, a  new test was added for the q_elimination sampling method for RB circuit construction.