Bugfixes for edge case in fiducial selection
Bug fixes related to edge cases revealed from updates to the tutorial notebooks. The first edge case had to do with changes to the generalized deduping routine when the construction of equivalence classes resulted in a single class. The second was due to the fact that the models in modelpacks use an ExplicitStateSpace for some reason, and this lacked a method/attribute similar to qudit_labels for QuditSpace objects. Added a new method to the base StateSpace class that travereses the tensor product blocks and concatenates a list of state space labels.