Updates germ_length_limit interpretation to avoid dropping a germ completely.
Previously, in GST circuit construction when a maximum length was greater
than the per-germ maximum length given by the `germ_length_limit` argument,
the circuits for that germ were always omitted.  This could be annoying,
however, if *only* large maximum lengths are given and a germ has a
max-length limit smaller than all the maximum lengths requested.  This
commit alters the behavior of the GST circuit construction routines to
include the circuits for such a germ that have the germ's repeated length
equals its max-length-limit (even though this limit is smaller than the first
maximum length requested).