Adds some extra logic for processing gauge-optimization suite object when running GST.
Adds logic that treats a gauge optimization suite object with `None` as its
suite names (checks of `gaugeopt_suite.gaugeopt_suite_names is None`) as specifying
that no gauge optimization should be performed.  Such an object is build if
gaugeopt_suite=None is passed when creating a GST protocol, and so we should
handle this case gracefully (previously this would result in an error).