Add support for non-standard POVMs and qudits
This patches the write_empty_protocol_data function to add in the previously missing support for non-standard POVMs and for standard qudit POVMs (i.e. with dimension >2). For qudits this also supports having qudits of different dimensions. Not included is anything for handling instruments, which will require some additional work if desired.