Model Violation Figures Corrections
Corrected the model violation overview figure and table (pushed earlier) to take into account Poisson bump to the effective degrees of freedom. Corrected the per sequence detail boxplot. The scatterplot is currently broken (colors are correct, but values no longer show correctly). This should be fixed soon. The histogram probably cannot be easily fixed and is showing the colors and line for the highest effective dof.