Enforce stricter compatibility when adding circuits
This commit adds a check to the add dunder method for Circuit objects to enforce compatibility across their respective line labels. In particular, a ValueError is now raised when adding circuits with placeholder default '*' line labels to those with standard specified labels. This is apparently something that by design we don't want to be possible, and the lack of this check has led to some unexpected, apparently buglike, behavior in the past. Also adds a docstring to this method explaining this requirement.