DensePureState serialization bugfix
This addresses a deserialization bug that occurred when deserializing certain DensePureState objects. I am not sure why, but in the original code the _basis attribute for these was in some cases intentionally set to None, which caused problems in the deserialization pipeline. Not sure what the intention was there (so maybe there are subtle consequences to this change), but simply keeping this set seems to have done the trick.