Update calc1Q comparison files and performance tweaks
Switch to using a minimally informationally complete set of fiducials to reduce test experiment size and also significantly increase number of shots in data set  and the magnitude of the noise which significantly improves the convergence time for the optimization. Updates comparison files in accordance with this change and in accordance with a bugfix related to serialization in a previous commit.