Speed up and modernize more tests (plus broken test fixes)
This commit speeds up and brings in line with modern code the time dependent GST unit test module. Also included are fixes and modernization of the report tests, and some related bugfixes in the actual source code. Minor change to which convolve function is used in drift library to address deprecation warning. Minor bugfixes to report generation code for plot we probably don't use all that often, but do test the generation of.