Create a new modelmember for affine shifts
This commit adds in a new modelmember called AffineShiftOp for modeling operations that correspond to purely affine shifts of a vector in Hilbert-Schmidt space. This is useful as a building block for the new ASMP representation being used in certain FOGI contexts. As part of this update the implementation of ProtectedArray has been significantly rewritten to allow a more general specification of protected indices
which is needed for restricting to the form of the affine matrices in question.