Add fallback gate name search
Add a fallback behavior which uses a new function for checking if a unitary corresponds to a built-in gate up to an overall phase to try and search for a matching gate if one is missing from the conversion dictionary. A warning is raised when this happens to alert the user.