Add better handling for GST w/o gauge optimization
This commit adds better handling for GST without gauge optimization with regards to report generation. This ports over some changes that were made earlier in the context of model testing, another setting in which not having model tests is commonplace. This is done by adding a new model to the estimate called 'trivial_gauge_opt' which is each to 'final iteration estimate' and when detected by the report generation gets subbed in for the generation of gauge dependent results. Also includes some docfixes and removal of debug commands.