Make LinearGateSetTomography serializable
I noticed that LinearGateSetTomography was missing auxfile information of the sort that StandardGST and GateSetTomography do for meta directory based serialization. This commit brings this protocol in line with the other two and enables directory based serialization for it. Also adds in a set of unit tests for testing the writing to directory and reading from directory of all three of the aforementioned classes.