Partially undo deprecation of 'none' suite name
It occurred to me in writing the PR for this change why there was an option for 'none' as a suite name in the first place, and that is because one can specify a list of different gauge-optimizations to perform, including the case of no gauge optimization. I am guessing, though, that there will probably wind up being a secondary edge case with broken reporting where if you do have a list of multiple gauge opt suites, the gauge variant figures for the 'none' entry will probably not generate properly.