Bugfixes for multiple gauge optimizations with none included
There was a bug present when using multiple gauge optimizations/suites when including 'none' as one of the suboptions wherein the model corresponding to 'none' was not included in the estimate. This bug fix includes a bunch of new edge case handling for when 'none' is included in a list of gauge optimization suites and a number of fixes related to this that ensure it properly shows up in the report.