Fixes bug in DenseOperator.kraus_operators when Choi matrix has degenerate spectrum.
Switches a numpy.linalg.eig to numpy.linalg.eigh call to ensure that
the matrix of eigenvectors is unitary (not always the case for eig,
in particular when acted-on matrix has a degenerate spectrum), and
adds an assertion statement to ensure eigh is working properly.  I think
the reason eig was used in the first place is that we've seen buggy
behavior of eigh, and so the assertion should keep it in check.

Adds unit test that verifies Kraus decomposition works as expected
for a depolarizing channel (as a dense op).