Per-sequence detail visualization controls
This commit adds a few new options for finer grained control of the per-sequence detail visualizations.
1. This adds intra-plaquette grids for easier reading of FPRed experiment designs. These can be toggled on and off with a button.
2. The option to toggle on and off hover labels on boxes.
3. A new visualization option that enables click-to-display of the hover label information in an area off to the side of the figures plot area. This can also be toggled on and off.

Additionally there is some clean up for long since out of data code. This also adds a copy of the open sans true type file to the repo as this seemed to be the most portable solution to the problem of having a font handy and readily findable by pillow for estimating the rendered size of the hover labels (needed to ensure we have enough margin included so that the text doesn't get cut off).