Performance enhancements for report generation
This commit adds in additional plumbing for the MDC stores generated during the course of GST estimation so that these can be accessed by the report generation, side-stepping the need to rebuild them from scratch when creating ColorBoxPlots. This only provides a small speedup for relatively small max Ls, but a fairly significant one for large max Ls. e.g. report generation for one-qubit and a maximum depth of 2048 is sped up by ~40% with these changes. Even longer max Ls should be sped up by an even larger fraction.