Attempt to migrate tests to reuseable Actions.
This may temporarily break the tests,
but the goal is to minimize maintenance burden
by having a single action that can run any combination of tests
(since the environments are the same between them),
and switch tests on and off as desired on various branches.
This ALSO tries to move towards using conda environment to solve our MacOS 14 + Python 3.8 + cvxopt problem.