Implement Faster Circuit Primitives
This commit speeds up the implementation of some core circuit primitives used in pyGSTi. The most notable are the circuit inequality check and hashing.

1. Equality checking is sped up by introducing a relatively inexpensive length comparison which allows fast short circuiting of the test.
2. For hashing we now leverage staticness more directly by caching both the hash, and the tuple representation used for hashing.

To go along with these changes are (pre-emptive) bugfixes for some inplace circuit modification methods which previously did not check for whether a circuit was static (which would result in hash invalidation). Similarly we have added restrictions to the setter methods for some circuit parameters that feed into the hash to prevent modification of these for static circuits.